// src/azure/benefits.d.ts
export type BenefitType = 'reservation' | 'savings-plan';
// ...
export interface IBenefitUtilizationSummary {
    total: number;
    withData: number;
    sevenDayAverage?: number;
    thirtyDayAverage?: number;
    byBenefitType: Array<{
        benefitType: BenefitType;
        total: number;
        withData: number;
        sevenDayAverage?: number;
        thirtyDayAverage?: number;
    }>;
}